Refactor sample app into clean feature packages
Split the 274-line BootUiSampleApplication god class into one-class-per-file
units and organize the flat io.github.jdubois.bootui.sample package into
feature sub-packages:

- catalog: Product, ProductRepository, ProductSummary, SampleCatalog,
  SampleController, SampleSettings
- chat: ChatController
- web: HelloController, AdminController
- scheduling: EchoScheduler
- config: SecurityConfiguration, NativeHintsConfiguration, and the extracted
  SampleMigrationConfiguration (Flyway/Liquibase demo beans)
- advisor.hibernate: the intentional Hibernate advisor demo entities/repos
- advisor.architecture: ArchitectureIssuesController

Behavior is unchanged: class names, HTTP paths, cache names, the scheduled
task, the sample.* config prefix, and the io.github.jdubois.bootui.sample
base package (logging) are all preserved, so every advisor finding and panel
keeps working. Updated the data.spec.js FQN assertion for the relocated
ProductRepository and added imports in the integration test.
