+++ bootui-sample-app/src/main/java/io/github/jdubois/bootui/sample/advisor/architecture/ArchitectureIssuesController.java
@@ -1,5 +1,6 @@
-package io.github.jdubois.bootui.sample;
+package io.github.jdubois.bootui.sample.advisor.architecture;
 
+import io.github.jdubois.bootui.sample.catalog.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
