+++ bootui-sample-app/src/main/java/io/github/jdubois/bootui/sample/advisor/hibernate/SampleInvoice.java
@@ -1,4 +1,4 @@
-package io.github.jdubois.bootui.sample;
+package io.github.jdubois.bootui.sample.advisor.hibernate;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
