+++ bootui-sample-app/src/main/java/io/github/jdubois/bootui/sample/advisor/hibernate/SampleOrderRepository.java
@@ -1,4 +1,4 @@
-package io.github.jdubois.bootui.sample;
+package io.github.jdubois.bootui.sample.advisor.hibernate;
 
 import java.util.Collection;
 import java.util.List;
