+++ bootui-sample-app/src/main/java/io/github/jdubois/bootui/sample/catalog/ProductSummary.java
@@ -1,0 +1,10 @@
+package io.github.jdubois.bootui.sample.catalog;
+
+import java.io.Serializable;
+
+public record ProductSummary(Long id, String name, String category, boolean active) implements Serializable {
+
+    static ProductSummary from(Product product) {
+        return new ProductSummary(product.getId(), product.getName(), product.getCategory(), product.isActive());
+    }
+}
