+++ bootui-sample-app/src/main/java/io/github/jdubois/bootui/sample/catalog/SampleSettings.java
@@ -1,0 +1,27 @@
+package io.github.jdubois.bootui.sample.catalog;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "sample")
+public class SampleSettings {
+
+    private String greeting = "Hello";
+
+    private int retries = 3;
+
+    public String getGreeting() {
+        return greeting;
+    }
+
+    public void setGreeting(String greeting) {
+        this.greeting = greeting;
+    }
+
+    public int getRetries() {
+        return retries;
+    }
+
+    public void setRetries(int retries) {
+        this.retries = retries;
+    }
+}
