+++ bootui-sample-app/src/main/java/io/github/jdubois/bootui/sample/config/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package io.github.jdubois.bootui.sample;
+package io.github.jdubois.bootui.sample.config;
 
 import static org.springframework.security.config.Customizer.withDefaults;
