+++ bootui-sample-app/src/main/java/io/github/jdubois/bootui/sample/web/AdminController.java
@@ -1,0 +1,15 @@
+package io.github.jdubois.bootui.sample.web;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/admin")
+public class AdminController {
+
+    @GetMapping
+    public String admin() {
+        return "BootUI sample admin";
+    }
+}
